refactor(solver): split mapped key/source extraction into a submodule to restore the file-size ratchet
`evaluation/evaluate_rules/mapped.rs` had grown to 1994 lines — 12 over its
per-file ratchet ceiling of 1982 — which left `test_solver_oversized_file_size_ceilings`
red in the solver direct-unit job. Two ratchet guards had been `#[ignore]`d
("file-size ratchet is currently red in the direct unit CI job") rather than
reconciled.

Extract the key-derivation half of mapped-type evaluation (building
`MappedKey`/`MappedKeys` from literals, properties, and `keyof` constraints,
plus homomorphic / template / `keyof` source recovery) into a new
`mapped/key_extraction.rs` submodule, mirroring the existing
`keyof_constraint` / `keys_guard` split in the same directory. The methods
move verbatim onto a second `impl TypeEvaluator` block with `pub(super)`
visibility so the `evaluate_mapped` entry point and the `keys_guard` wrapper
keep calling them unchanged. No evaluation logic changes.

- `mapped.rs`: 1994 -> 1375 lines; new `key_extraction.rs`: 639 lines.
- Tighten the `solver_file_eval_mapped` baseline 1982 -> 1375 (ratchet down).
- Un-ignore `test_binder_file_size_ceiling` (passes now that the job is green).

Behavior is unchanged: the full `tsz-solver --lib` mapped suite (406 tests)
passes, and all eight file-size ceiling guards pass with zero ignored.
