+++ crates/tsz-solver/src/tests/file_size_baselines.txt
@@ -27,7 +27,7 @@ solver_file_narrowing_core 2718
 solver_file_relations_compat 2950
 solver_file_constraints_walker 2348
 solver_file_diag_format_mod 2318
-solver_file_eval_mapped 1982
+solver_file_eval_mapped 1375
 solver_file_subtype_generics 2217
 solver_file_call_args 2097
 solver_file_eval_index_access 2226
